backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic==2.9.2
langchain
langchain-openai
langchain-community
redis
chromadb
tiktoken
requests
